Switch dealer::player_container from std::vector to std::array
Copy/move operations on round class had to be deleted because this
structure holds pointers to its own internals. Default copying leads to
unwanted and very error-prone semantics. Custom copying makes no sense
since the resource the class operates on is unique.

As an alternative to `_betting_round = {...}`, we use placement new.