Problem: pot misbehaves when no bets are placed.
It appends all the players to _eligible_players, including the ones who
are already in.

Solution: clear the _eligible_players before re-adding all the players.
As explained in the comments, it is possible that a player folds on a
no-bet betting round. This is why instead of no-op, we re-add all the
indices over again.
Some additional comments have also been added to aid in understanding
this function.
Regression test is included.