Problem: Contract check too strong.
table::betting_rounds_completed contract check requires for betting
round to be in progress. However, it does make sense to check if betting
rounds are completed at any point in the hand. In this form, the
function is not useful for the user because if the betting rounds is not
in progress after ending it, there is no need to call this function (it
can be deduced).

Solution: Weaken the contract to require only the hand to be in
progress when calling it.