Problem: Cannot set the table button manually.
Solution: Allow the user to set the button at the start of every hand.
It cannot be set at any time because we would have to keep track of
whether the button is on the existing player. Since the button is also
set manually on the first hand of the table, rename the
_first_time_button boolean to _button_set_manually, so the name is more
general.