Problem: hole cards exist longer than the hand.
This creates windows of time where the player.hole_cards property is
semantically uninitialized (enum values set to 0). This is a flaw in the
interface. It leaves room for the user to make errors. It also
propagates this bad design to the higher levels.

Solution: Remove the player.hole_cards property and replace it with a
std::array<hole_cards, max_players> inside the dealer (stored for the
length of the hand).