Problem: crash when player to act stands up.
This happens in the case where only one other player is left remaining.
When the player to act stands up, the action is taken directly (instead
of being queued as an automatic action) and the hand ends naturally and
by itself.
However, we are calling act_passively() even in this case, which is
wrong since there is no more actions to be taken, and thus we violate
the contract.

Solution: Rearrange the control flow to do the right thing and add the
regression test.