Problem: going all-in preflop induces an automatic call.
This is a problem with how we evaluate
table::single_active_player_remaining. In this case, there is only one
player left in dealer::players (indirectly round::players) and that's
why act_passively() triggers. Using dealer::players is not correct.

Solution: Expose dealer::betting_round_players, that is, players who
started the current betting round *and* have not folded (folding players
are discarded manually internally to help with pot evaluation). On this
condition, we only need to require that player has not stood up, and
that's how we get the real correct definition of "active player" in the
context of the table class.