Problem: legal_actions raise range starts from 0.
You obviously cannot raise [0, bb) and returning that to the user is
giving misleading and wrong information.

In the context of betting_round, we are treating min_raise and
biggest_bet as the same thing in the beginning. But this is wrong. It is
true only for the preflop. For instance, on the flop, the biggest_bet
should indeed be 0, but min_raise is not 0, it's the big blind (or
anything specific with min_raise, not necessarily the big blind).

Solution: Provide another argument to the betting_round constructor,
taking biggest_bet. It is 0 by default because you should only set it
for the preflop. Fix the constructor uses in the dealer class, and fix
the betting_round tests to have the behavior equivalent to the former.