Problem: cannot tell all-in and folded players apart.
class round only cares about whether the player is active or inactive
(meaning can he take further action or not). However, users of this
class have good reasons to care.

Solution: Instead of using a boolean value, we will introduce an
enumeration player_state with three distinct states that a player can be
in: inactive, idle and active.
- inactive -- player has folded;
- idle -- player went all-in;
- active -- player has chips left behind, and can take potential further
action.